jugador completado, juego falta la funcion añadir

+++ SpringBootApp/src/main/java/com/zubiri/app/Interfaces/DBJuegosInterface.java
@@ -1,11 +1,10 @@
 package com.zubiri.app.Interfaces;
 
-import java.util.ArrayList;
 
 import org.springframework.data.repository.CrudRepository;
 
 import com.zubiri.app.beans.Juego;
 
-public interface DBJuegosInterface extends CrudRepository<Juego, Long>{
+public interface DBJuegosInterface extends CrudRepository<Juego, Integer>{
 	
 }
